modularized employeeController
