+++ server/Controllers/employeeController/middleware/deleteEmployee.js
@@ -8,3 +8,30 @@
  *
  * **************************************************
  */
+
+require('dotenv').config();
+const path = require('path');
+const db = require(path.resolve(__dirname, '../../../models/employeeModel'));
+
+const deleteEmployee = async (req, res, next) => {
+  const { _id } = res.locals.targetEmployee;
+  try {
+    const input = [];
+    let employeeQuery = 'DELETE FROM employee ';
+   if(_id){
+     employeeQuery += `WHERE _id = $1 RETURNING * `;
+     input.push(_id);
+   }
+    const deleted = await db.query(employeeQuery, input);
+    res.locals.deleted = deleted.rows[0];
+    return next();
+  }
+  catch(err){
+    return next({
+      log: 'employeeController.deleteEmployee: ERROR',
+      message: {err: 'Error occurred in employeeController.deleteEmployee'}
+    });
+  }
+};
+
+module.exports = deleteEmployee;
